Seperated context into two files; Fixed the Gallery Screen's Progress bar's size;
