+++ client/App.js
@@ -3,13 +3,16 @@ import 'react-native-gesture-handler';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import MainNavigator from './src/navigation/MainNavigator';
 import { PermissionProvider } from './src/context/PermissionsContext';
+import { ImageProvider } from './src/context/ImageContext';
 
 export default function App() {
   return (
     <PermissionProvider>
-      <SafeAreaProvider>
-        <MainNavigator />
-      </SafeAreaProvider>
+      <ImageProvider>
+        <SafeAreaProvider>
+          <MainNavigator />
+        </SafeAreaProvider>
+      </ImageProvider>
     </PermissionProvider>
   );
 }
